Add test for omitted thresholds
Current behaviour is good if not a little confusing.  The order of the
(warn, crit) and (min, max) thresholds are reversed between the API and
the output format documented in the plugin guidelines [1]:

  output:        [warn];[crit];[min];[max]
  nagiosplugin:  min, max, warn, crit (collectively vararg'd)

This test helped me confirm that all was well.  Existing tests do not cover
this particular edge case.

1: https://nagios-plugins.org/doc/guidelines.html#AEN200